Move envelope space charge transfer matrix calculations
Move to Envelope class
